Adjusted code so only one class is accessing/modding database.
Now app.js will act as a main router for dApp users to make API calls to and it will route to either Layer 1, 2, or 3. But Layer Zero (located in middleware) will deal with the MySQL code. This makes it easier in the future for when GVM is being built, only this middleware class needs to be modified.